LayoutTests/crypto/subtle/ecdh-import-jwk-private-key-p256-expected.txt

Test importing a JWK ECDH private key with P-256

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.extractable is false
PASS privateKey.algorithm.name is 'ECDH'
PASS privateKey.algorithm.namedCurve is 'P-256'
PASS privateKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
